requirements.txt
streamlit==1.51.0
tensorflow==2.20.0
numpy==1.26.4
pandas==2.3.3
pillow==12.0.0
scipy==1.16.3
